erc20 testing and offer acceptance

+++ contracts/JPToken.sol
@@ -4,7 +4,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract JPToken is ERC20, Ownable{
-    constructor(uint initialSupply)
+    constructor(uint256 initialSupply)
     ERC20("JPTOKEN20NAME", "JPTOKEN20SYM")
     {
         _mint(msg.sender, initialSupply);
